Add files via upload
Email is a critical means of communication in today’s world, but it can be time consuming and tiring task to send out emails manually. Fortunately, Python which I am very passionate and proud of, provides a simple and efficient way to automate email sending and receiving tasks. 
This simple Email Automation project is a beginner-friendly Email Automation system with Python. It helps make sending emails easier and more efficient by adding useful features. With this system, you can:
* Send both plain-text emails and emails with fancy formatting Hyper Text Markup Language (HTML). 
* Attach files like PDFs to share important documents.
* Uses SSL encryption (smtplib.SMTP_SSL) to ensure a secure transmission of emails.  Here, I am using Python's built-in library called smtplib. 
* Simple Mail Transfer Protocol (SMTP)